requirements.txt
Flask==2.3.3
face-recognition==1.3.0
Pillow==10.0.1
numpy==1.24.3
cmake==3.27.7
dlib==19.24.2